+++ tests/HasAppTest.php
@@ -32,14 +32,6 @@ class HasAppTest extends PHPUnit_Framework_TestCase
         $class2->setApp($app2);
         $this->assertTrue($class->getApp() !== $class2->getApp());
     }
-
-    public function testInjectAppDeprecated()
-    {
-        $app = Mockery::mock('Infuse\Application');
-        $class = new SomeClass();
-        $this->assertEquals($class, $class->injectApp($app));
-        $this->assertEquals($app, $class->getApp());
-    }
 }
 
 class SomeClass